fix the 3g interface configuration
The first problem is the "chat" command which misses "-E", so the
environment variable addressed inside the script is never used and
the script breaks.

The second problem is the that you cannot assume an Option card when
you did "not discover a Novatel".
In my case the script tried to set the mode with the Option
AT-command which failed on the Huawei stick.
Unfortunaltely I don't have a correct command for that device, so I
just want to start the connection without any mode set before thus using
the defaults.

Signed-off-by: Markus Wigge <mwigge@marcant.net>


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@10774 3c298f89-4303-0410-b956-a3cf2f4a3e73