Hi,
this is the first of three patches to fix the ethernet driver of RDC R32xx applied to svn r10754.

It adds debugs and make some procedures (r6040_*) global symbols in order to help with debugging kernel oops. Version is increased from 0.16 to 0.18 to avoid conflicts with the driver variant metioned in the X-WRT forum (call 0.17.3) Tested on a Sitecom WL-153.

Signed-Off-By: Joerg Albert <jal2@gmx.de>


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@11506 3c298f89-4303-0410-b956-a3cf2f4a3e73